requirements.txt
aio-pika==9.1.4
aiormq==6.7.6
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
cffi==1.15.1
click==8.1.4
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.100.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
install==1.3.5
iso8601==1.1.0
multidict==6.0.4
pamqp==3.2.1
priority==2.0.0
pycparser==2.21
pydantic==1.10.11
pydantic_core==2.1.2
pypika-tortoise==0.1.6
pytz==2023.3
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
wsproto==1.2.0
yarl==1.9.2
